chore: rename tablecodegen to databasecodegen
More descriptive name.
