+++ internal/codegen/databasecodegen/database_test.go
@@ -1,4 +1,4 @@
-package tablecodegen
+package databasecodegen
 
 import (
 	"testing"
