+++ internal/codegen/databasecodegen/goldenfile_test.go
@@ -1,4 +1,4 @@
-package tablecodegen
+package databasecodegen
 
 import (
 	"fmt"
