+++ internal/codegen/databasecodegen/interleavedreadtransaction.go
@@ -1,4 +1,4 @@
-package tablecodegen
+package databasecodegen
 
 import (
 	"strconv"
