+++ internal/codegen/databasecodegen/interleavedrow.go
@@ -1,4 +1,4 @@
-package tablecodegen
+package databasecodegen
 
 import (
 	"fmt"
