+++ internal/codegen/databasecodegen/row.go
@@ -1,4 +1,4 @@
-package tablecodegen
+package databasecodegen
 
 import (
 	"fmt"
